ghc: Add `enableDwarf` flag, on by default.
This allows our GHCs to build programs with DWARF
debug information when `-g` is passed, see
https://ghc.haskell.org/trac/ghc/wiki/DWARF.

This can make debugging of many Haskell bugs and
performance issue significantly easier.

This change is expected to increase the size of
GHC and Haskell libraries, as debugging symbols are
added to the object files.
The size of the eventual Haskell binaries that are not
explicitly built with `ghc -g` should remain
unaffected when the standard statish linking is used
(that links Haskell libraries statically, and system
libraries dynamically, which is nixpkgs' default).