fetchurl: Don't force-override curl's gssSupport to on. Fixes #66499.
The original intent in commit

    a1fec866 treewide: assemble all `fetchurlBoot` uses in overrides to `fetchurl` itself

was to turn `gssSupport` *off* on Darwin, but the code actually also forced
it *on* on Linux.
This resulted in previous (e.g. overlays) `.override { gssSupport = false; }`
being ignored (#66499).

This commit fixes it by just respecting the old value when it doesn't need
to be forced to off.