ceph: arrow 19 compilation fix with protobuf 30
It seems that Arrow 20 which broke compilation initially causes issues with our tests when the patch which was merged upstream is applied.
At the same time the bundled Arrow 17 can't work with current Re2 due to its C++ version.
The thing that broke the Arrow 19 build is because Arrow 19 is itself not compatible with Protobuf 30, which is pulled in somewhere along the dependency graph by something else causing linking issues.
However, the patch for Protobuf 30 compat applies cleanly to our Arrow 19, which means we can dump the older version pin and instead use this patch to get Arrow 19 working again.

Signed-off-by: benaryorg <binary@benary.org>